Add sun.ni.ch module opens to getting started docs
As noted in [this issue](https://github.com/uncomplicate/neanderthal/issues/115#issuecomment-900707935), we need to also open
the `java.base/sun.nio.ch` module to get access to the `DirectByteBuffer Cleaner`. Since we can't enable this by default and retain support for Java 8, users of Java 9+ need to set this up by hand.